Move out `omega` files to `account` FIX #39
Massive tree re-organisation.
Purpose : a project tree respecting more the Django habits.

Project `omega` :
* keep errors & base templates, move all others
* keep all statics
* exports all tests & views
* disabling admin interface
* rename `/my/*` routes

App `account` :
* rename some routes
* host most of ex-`omegega` templates
* no `urls.py` : no need to prefix URL with `account`

App `ersatz` :
* host `searchform.html`
* remove unused `home.html`
