+++ account/templates/registration/password_change_form.html
