+++ account/templates/registration/password_reset_form.html
