+++ account/tests.py
@@ -1,3 +1,10 @@
-from django.test import TestCase
+import pytest
 
-# Create your tests here.
+import account.forms
+
+
+def test_account_forms_create_user_fields():
+    test_form = account.forms.SignInForm()
+
+    for field in ('username','email','password1','password2'):
+        assert field in test_form.fields
