+++ account/views.py
@@ -1,4 +1,28 @@
-from django.shortcuts import render
+from django.shortcuts import render, redirect
+from django.contrib.auth import login, authenticate
 
-def index(request):
-    return render(request, 'account/home.html')
+from account.forms import SignInForm
+
+
+def home(request):
+    return render(request, 'home.html')
+
+def about(request):
+    return render(request, 'about.html')
+
+def account(request):
+    return render(request, 'account/account.html')
+
+def signin(request):
+    if request.method == 'POST':
+        form = SignInForm(request.POST)
+        if form.is_valid():
+            form.save()
+            username = form.cleaned_data.get('username')
+            raw_password = form.cleaned_data.get('password1')
+            user = authenticate(username=username, password=raw_password)
+            login(request, user)
+            return redirect(request.POST['next'])
+    else:
+        form = SignInForm()
+    return render(request, 'registration/signin.html', {'form': form})
