--- ersatz/templates/ersatz/home.html
@@ -1,16 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %} - ersatz app index{% endblock %}
-
-{% block content %}
-  <h2>ersatz</h2>
-
-  {% if user.is_authenticated %}
-  <p>Dear <strong>{{ user.get_full_name }}</strong> , here's the <strong>ersatz app index</strong>.</p>
-  <p>Try to <a href="{% url 'ersatz-search' %}?foo=bar&s=sel" >search something</a></p>
-
-  {% else %}
-  <p>Hello <strong>world</strong>, here's the <strong>ersatz app index</strong>.</p>
-  {% endif %}
-
-{% endblock %}
