+++ ersatz/templates/ersatz/list.html
@@ -13,7 +13,7 @@
     {% if not favorites %}
                 Vous n'avez pas encore enregistré de favoris, il est temps de faire une petite recherche!
             </h4>
-                {% include 'omega/searchform.html' %}
+                {% include 'ersatz/searchform.html' %}
     {% endif %}
           </div>
         </div>
