+++ omega/templates/base.html
@@ -44,11 +44,11 @@
         <div class="collapse navbar-collapse" id="navbarResponsive">
           <ul class="navbar-nav ml-auto">
             <li class="nav-item">
-              {% include "omega/searchform.html" %}
+              {% include "ersatz/searchform.html" %}
             </li>
           {% if user.is_authenticated %}
             <li class="nav-item">
-              <a class="nav-link js-scroll-trigger" href="{% url 'account-index' %}" title="Mon compte"><i class="fas fa-user fa-2x mb-3" title="Mon compte"></i></a>
+              <a class="nav-link js-scroll-trigger" href="{% url 'account' %}" title="Mon compte"><i class="fas fa-user fa-2x mb-3" title="Mon compte"></i></a>
             </li>
             <li class="nav-item">
               <a class="nav-link js-scroll-trigger" href="{% url 'ersatz-list' %}" title="Mes produits">
@@ -82,7 +82,7 @@
           </div>
           <div class="col-lg-4 mx-auto">
             {% block headform %}
-              {% include "omega/searchform.html" %}
+              {% include "ersatz/searchform.html" %}
             {% endblock %}
           </div>
         </div>
