+++ omega/urls.py
@@ -1,17 +1,15 @@
-"""
-omega URL Configuration
-"""
 from django.contrib import admin
 from django.urls import include, path
 
-from omega import views as omega_views
+from account import views as account_views
 
 urlpatterns = [
-    path('', omega_views.index, name='home'),
-    path('my/signin/', omega_views.signin, name='signin'),
-    path('my/', include('django.contrib.auth.urls')),
-    path('account/', include('account.urls')),
-    path('about/', omega_views.about, name='about'),
+    path('', account_views.home, name='home'),
+    path('about/', account_views.about, name='about'),
+    path('account/', account_views.account, name='account'),
+    path('account/signin/', account_views.signin, name='signin'),
+    # path('admin/', admin.site.urls),
+    path('auth/', include('django.contrib.auth.urls')),
     path('ersatz/', include('ersatz.urls')),
-    path('admin/', admin.site.urls),
+
 ]
